requirements.txt
pytest==8.4.2
playwright==1.49.0
allure-pytest==2.13.5
pytest-xdist==3.8.0
pytest-rerunfailures==13.0
pytest-playwright==0.6.2